clone: Just invoke git clone directly; don't do anything extra anymore
git checkout DWIMery for 'git checkout <non-existent-branch>' obsoletes
the main purpose behind the special clone effort.  Besides, that code was
buggy anyway for a repository that had multiple branches whose basename
was the same, e.g. a repository with both of the branches
  elem-loops-test
  collab/aria/elem-loops-test
would not be cloned well by e.g. clone as it was attempting to set up two
local branches of the same name to track the two different remote tracking
branches.  I could have fixed that, but the automatic branch setup just
isn't useful anymore.