status: Add support for unusual combination of flags from git status
status has -s (short for --short) and -u (short for --untracked) flags.
The -u flag takes one of three strings -- no, normal, or all.  That
string need not be separated from -u with an equal sign or a space.
But what makes this really odd, is that -s and -u can be run together
as well, giving us commands like
   git status -suno
which means
   git status -s -u no
However, forms like
   git status -unos
are not allowed; it is treated like
   git status -u nos
which makes no sense.

I find this combination a little odd, but since eg is supposed to support
essentially everything that git does and we try to only differ in
limited and explicitly enumerated and documented cases, add support for
this particular combination of command line arguments.