Release 1.7.5.2
I was going to create a 1.7.4 release, since git-1.7.4 was the version
with the infamous status output change that broke eg (no other version of
git appears to have introduced problems with eg yet), but one of the eg
specific tests in the testsuite triggered a bug in git that was not fixed
until 1.7.5.2.  If only I had been a bit more prompt at trying at trying
to update eg, I could have caught the bug and sent it upstream.  It was
really lame that I fixed the bug locally and never got that fix pushed
upstream.  Looks like at least a few others easily figured it out too,
though.  Oh well.

This release also includes a number of merge requests from other folks,
some of which had sat for years.  I'm such a horrible maintainer.  But,
this time I updated the HACKING file to try to make it easier for others
to know how to maintain it, and I hope that I'll have a little more time
to keep it up to date.