push: Use git's push.default setting to simplify our push implementation
We did a lot of work to determine the default remote to push to and to
determine the upstream tracking branch so that we could craft an
appropriate command line for git push.  Those rules change over time,
though (e.g. remote.pushdefault and branch.<name>.pushremote) which makes
this a pain.  Besides, git has for a while had a push.default setting
which we can use to avoid most this logic and let git do the correct thing
for us.