Slight tweak to determination of new unknown files
Apparently, if there is a subdirectory which itself only contains a .git
directory, the --no-empty-directory flag to ls-files was returning that
directory in some cases but not in others.  Simple workaround was to not
use that flag when recording ignored unknown files, but to continue using
it when checking whether there were new unknowns that we wanted to warn
about.  This fixes the testcase in git-1.8.5 that was causing me grief
without regressing adb7312 (Make the list of new unknowns a subset of the
output of eg status, 2008-05-12)