Improve the check for whether we have no commits in the repo
Sadly, this check was named "initial_commit", which meant that the check
was obviously implemented by determining if HEAD was a branch that was yet
to be born.  Replace this with the correct check, namely whether show-ref
returns 0 lines of output.

I should probably also rename this function to match its intended purpose.
*shrug*