Change 'unknown'->'untracked' and fix error message when > 5 new-unknowns
Since we are now just using 'git status' as-is, people will get used to
the terminology of 'untracked'.  So, we may as well use it everywhere.
Also, since we can no longer tell people to run 'eg status' to get a full
list of new untracked files, make a separate command for just getting the
new untracked files and refer to it in the error message that only prints
the first five new untracked files.