bash-completion-eg.sh: Update to cool new stuff
git's bash completion script has had a number of updates.  Update
bash-completion-eg.sh to work better with these new versions, take
advantage of new capabilities to simplify maintenance, and add support
for a few more flags.

Newer git also has some testcases for bash completion, which naturally I
will be co-opting as testcases for eg's bash completion similar to how I
handle the rest of the git testsuite as a testsuite for eg.  That should
make it easier to notice breakages in eg's bash completion and keep it up
to date if needed.